Remove the need to update wallet restoration information.
The only update we were left with was updating the `cancel` action,
but we can even do away with that by making `cancel` just read the
`Async` to cancel out of an MVar. When that MVar is empty, there is
nothing left to cancel. This will also make it easier to start new
restoration tasks after a fork: the `cancel` action is unchanged, we
just stick a new `Async` into the MVar and call it a day.